# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ctrlX System Identification API",
...
    "contact": {
      "name": "Equipo ctrlX SysID"
    },
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/opcua/discover-programs": {
      "post": {
        "tags": [
          "OPC UA"
        ],
        "summary": "Listar los programas PLC del dispositivo",
        "description": "Conecta con las credenciales dadas y devuelve los programas colgados de\n`Objects/Datalayer/plc/app/Application/sym`.\n\nEs el **paso 2** del flujo: sirve para llenar el desplegable de programas.",
        "operationId": "discover_programs_api_opcua_discover_programs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpcUaDiscoverProgramsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverProgramsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Faltan datos o son inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__schemas__opcua__ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Fallo al hablar con el PLC",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__schemas__opcua__ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/opcua/discover-variables": {
      "post": {
        "tags": [
          "OPC UA"
        ],
        "summary": "Listar TODAS las variables de un programa",
        "description": "Devuelve todas las variables del programa con su tipo y su valor actual,\n**sin filtrar por nombre**. La vista decide cuál es actuador, sensor,\nsetpoint, tiempo y tipo de señal.\n\n`suggested_mapping` trae una propuesta basada en los nombres clásicos\n(`rActuator`, `rSensor`, ...) para pre-seleccionar los desplegables. Es\nsolo una sugerencia: si el programa usa otros nombres, esos roles vienen\nen `null` y los elige el usuario.\n\nEs el **paso 3** del flujo.",
        "operationId": "discover_variables_api_opcua_discover_variables_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpcUaDiscoverVariablesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverVariablesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Faltan datos o son inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__schemas__opcua__ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Fallo al hablar con el PLC",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__schemas__opcua__ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/opcua/login": {
      "post": {
        "tags": [
          "OPC UA"
        ],
        "summary": "Abrir sesión y arrancar la lectura cíclica",
        "description": "Valida la conexión, guarda el mapeo y arranca el hilo que lee el PLC\ncada 200 ms. A partir de aquí el WebSocket empieza a emitir `sample`.\n\nLimpia el buffer y cualquier identificación previa.\n\nEl campo `mapping` es opcional; los roles que se omitan se resuelven con\nlos alias por defecto. `mapping` en la respuesta es el **mapeo efectivo**,\nya completado.\n\nEs el **paso 4** del flujo.",
        "operationId": "opcua_login_api_opcua_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpcUaLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Faltan datos o son inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__schemas__opcua__ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Fallo al hablar con el PLC",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__schemas__opcua__ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiscoverProgramsResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "user": {
            "type": "string",
            "title": "User"
          },
          "programs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Programs",
            "description": "Programas colgados del nodo 'sym'"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "url",
          "user",
          "programs"
        ],
        "title": "DiscoverProgramsResponse",
        "example": {
          "ok": true,
          "programs": [
            "PLC_PRG",
            "PRG_Horno"
          ],
          "url": "opc.tcp://192.168.1.1:4840",
          "user": "boschrexroth"
        }
      },
      "DiscoverVariablesResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "user": {
            "type": "string",
            "title": "User"
          },
          "program_name": {
            "type": "string",
            "title": "Program Name"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles",
            "description": "Roles que espera el backend"
          },
          "variables": {
            "items": {
              "$ref": "#/components/schemas/PlcVariable"
            },
            "type": "array",
            "title": "Variables"
          },
          "suggested_mapping": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Suggested Mapping"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "url",
          "user",
          "program_name",
          "roles",
          "variables",
          "suggested_mapping"
        ],
        "title": "DiscoverVariablesResponse",
        "description": "Todas las variables del programa, sin filtrar por nombre.\n\n`suggested_mapping` es solo una sugerencia para pre-seleccionar los\ndesplegables: la vista decide el mapeo final.",
        "example": {
          "ok": true,
          "program_name": "PLC_PRG",
          "roles": [
            "time",
            "actuator",
            "sensor",
            "setpoint",
            "signal_type"
          ],
          "suggested_mapping": {
            "actuator": "rActuator",
            "sensor": "rSensor",
            "setpoint": "rSetPoint",
            "time": "rTimeSec"
          },
          "url": "opc.tcp://192.168.1.1:4840",
          "user": "boschrexroth",
          "variables": [
            {
              "data_type": "Float",
              "name": "rActuator",
              "node_id": "ns=2;s=.../rActuator",
              "numeric": true,
              "readable": true,
              "value": 12.0
            },
            {
              "data_type": "Float",
              "name": "rSensor",
              "node_id": "ns=2;s=.../rSensor",
              "numeric": true,
              "readable": true,
              "value": 7.21
            }
          ]
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "user": {
            "type": "string",
            "title": "User"
          },
          "program_name": {
            "type": "string",
            "title": "Program Name"
          },
          "mapping": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Mapping",
            "description": "Mapeo efectivo tras completar con alias los roles vacíos"
          },
          "started": {
            "type": "boolean",
            "title": "Started",
            "description": "True si el hilo de lectura cíclica arrancó"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "url",
          "user",
          "program_name",
          "mapping",
          "started"
        ],
        "title": "LoginResponse",
        "example": {
          "mapping": {
            "actuator": "rActuator",
            "sensor": "rSensor",
            "setpoint": "rSetPoint",
            "signal_type": "uiSignalType",
            "time": "rTimeSec"
          },
          "ok": true,
          "program_name": "PLC_PRG",
          "started": true,
          "url": "opc.tcp://192.168.1.1:4840",
          "user": "boschrexroth"
        }
      },
...
      "OpcUaDiscoverProgramsRequest": {
        "properties": {
          "user": {
            "type": "string",
            "title": "User",
            "description": "Usuario OPC UA del ctrlX"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "Contraseña OPC UA"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Endpoint OPC UA, formato opc.tcp://host:puerto"
          }
        },
        "type": "object",
        "required": [
          "user",
          "password",
          "url"
        ],
        "title": "OpcUaDiscoverProgramsRequest",
        "example": {
          "password": "********",
          "url": "opc.tcp://192.168.1.1:4840",
          "user": "boschrexroth"
        }
      },
      "OpcUaDiscoverVariablesRequest": {
        "properties": {
          "user": {
            "type": "string",
            "title": "User",
            "description": "Usuario OPC UA del ctrlX"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "Contraseña OPC UA"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Endpoint OPC UA, formato opc.tcp://host:puerto"
          },
          "program_name": {
            "type": "string",
            "title": "Program Name",
            "description": "Programa PLC devuelto por /discover-programs"
          }
        },
        "type": "object",
        "required": [
          "user",
          "password",
          "url",
          "program_name"
        ],
        "title": "OpcUaDiscoverVariablesRequest",
        "example": {
          "password": "********",
          "program_name": "PLC_PRG",
          "url": "opc.tcp://192.168.1.1:4840",
          "user": "boschrexroth"
        }
      },
      "OpcUaLoginRequest": {
        "properties": {
          "user": {
            "type": "string",
            "title": "User",
            "description": "Usuario OPC UA del ctrlX"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "Contraseña OPC UA"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Endpoint OPC UA, formato opc.tcp://host:puerto"
          },
          "program_name": {
            "type": "string",
            "title": "Program Name",
            "description": "Programa PLC a monitorear"
          },
          "mapping": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SignalMapping"
              },
              {
                "type": "null"
              }
            ],
            "description": "Mapeo elegido en la vista. Si se omite se usan los alias por defecto."
          }
        },
        "type": "object",
        "required": [
          "user",
          "password",
          "url",
          "program_name"
        ],
        "title": "OpcUaLoginRequest",
        "example": {
          "mapping": {
            "actuator": "Valvula_Salida",
            "sensor": "Temp_Horno",
            "setpoint": "Consigna",
            "time": "Tiempo_Proceso"
          },
          "password": "********",
          "program_name": "PLC_PRG",
          "url": "opc.tcp://192.168.1.1:4840",
          "user": "boschrexroth"
        }
      },
...
      "PlcVariable": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Browse name en el PLC"
          },
          "node_id": {
            "type": "string",
            "title": "Node Id",
            "description": "NodeId OPC UA"
          },
          "data_type": {
            "type": "string",
            "title": "Data Type",
            "description": "Tipo OPC UA: Float, Double, Int16... o UNKNOWN"
          },
          "value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Value",
            "description": "Valor en el momento de la lectura"
          },
          "numeric": {
            "type": "boolean",
            "title": "Numeric",
            "description": "True si sirve como señal continua para identificar"
          },
          "readable": {
            "type": "boolean",
            "title": "Readable",
            "description": "False si la lectura falló"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Motivo del fallo de lectura, si lo hubo"
          }
        },
        "type": "object",
        "required": [
          "name",
          "node_id",
          "data_type",
          "numeric",
          "readable"
        ],
        "title": "PlcVariable",
        "description": "Una variable del programa PLC.",
        "example": {
          "data_type": "Float",
          "name": "rSensor",
          "node_id": "ns=2;s=plc/app/Application/sym/PLC_PRG/rSensor",
          "numeric": true,
          "readable": true,
          "value": 7.21
        }
      },
...
      "SignalMapping": {
        "properties": {
          "time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time",
            "description": "Variable con el tiempo del proceso, en segundos"
          },
          "actuator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actuator",
            "description": "Variable de la salida del controlador (MV)"
          },
          "sensor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensor",
            "description": "Variable de la medición del proceso (PV)"
          },
          "setpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Setpoint",
            "description": "Variable del setpoint (opcional)"
          },
          "signal_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Type",
            "description": "Variable que indica la escala. Si vale 1 el backend convierte de 4-20 mA a %."
          }
        },
        "type": "object",
        "title": "SignalMapping",
        "description": "Qué variable del PLC cumple cada rol.\n\nTodos los campos son opcionales: el rol que se deje en `null` se resuelve\nen el backend con los alias por defecto (rActuator, rSensor, rTimeSec...).",
        "example": {
          "actuator": "rActuator",
          "sensor": "rSensor",
          "setpoint": "rSetPoint",
          "signal_type": "uiSignalType",
          "time": "rTimeSec"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "api__schemas__opcua__ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Formato de error de FastAPI.",
        "example": {
          "detail": "Falta la URL OPC UA."
        }
      },
...
    }
  },
...
}